refactor: enhance IPC handlers and improve workspace context resolution
- Replaced `ipcMain.handle` with `safeIpcHandle` in `healthCheck.ts` for better error handling.
- Updated type definitions in `healthCheck.ts` to use `LLMConfig`.
- Introduced `resolveWorkspaceFromEvent` function in `workspaceContext.ts` to streamline workspace resolution logic.
- Refactored file operations and terminal handlers in `index.ts` to utilize the new workspace resolution method.
- Added tests for `resolveWorkspaceFromEvent` to ensure correct functionality.
